% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexport.R, R/dataset_export.R
\docType{data}
\name{flat_titanic}
\alias{flat_titanic}
\title{Summaries of people and outcomes on titanic}
\format{
A data frame with 32 rows and 5 variables:
\describe{
\item{age}{price, in US dollars}
\item{class}{weight of the diamond, in carats}
\item{freq}{weight of the diamond, in carats}
\item{sex}{weight of the diamond, in carats}
\item{survived}{weight of the diamond, in carats}
...
}

A data frame with 32 rows and 5 variables:
\describe{
\item{class}{1st, 2nd, 3rd for each passenger}
\item{sex}{Male, Female for each passenger}
\item{age}{Child, Adult for each passenger}
\item{survived}{No, Yes for each passenger}
\item{freq}{count of individuals}
...
}
}
\source{
\url{http://www.diamondse.info/}

A version of the Titanic data from the datasets package
}
\usage{
flat_titanic

flat_titanic
}
\description{
A dataset containing blahhhhhhh

A dataset with a row for each combination of characteristics of individuals on the Titanic.
}
\keyword{datasets}
