% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexport.R
\docType{data}
\name{passengers}
\alias{passengers}
\title{Records for individual passengers on titanic; names recorded.}
\format{
A data frame with 1313 rows and 5 variables:
\describe{
\item{name}{weight of the diamond, in carats}
\item{class}{weight of the diamond, in carats}
\item{age}{price, in US dollars}
\item{sex}{weight of the diamond, in carats}
\item{survived}{weight of the diamond, in carats}
...
}
}
\source{
R stat
}
\usage{
passengers
}
\description{
A dataset ... blah blah
}
\keyword{datasets}
