% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{tidy_titanic}
\alias{tidy_titanic}
\title{A dataset titanic passengers and characteristics based on Titanic data from base R.  Includes Crew}
\format{
A data frame with 2201 rows and 5 variables:
\describe{
\item{id}{an id for each passenger}
\item{class}{class (1st, 2nd, 3rd or Crew)}
\item{sex}{sex of passenger}
\item{age}{child or adult}
\item{survived}{survival outcome of passanger}
...
}
}
\source{
R stats
}
\usage{
tidy_titanic
}
\description{
A dataset titanic passengers and characteristics based on Titanic data from base R.  Includes Crew
}
\keyword{datasets}
