% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexport.R
\docType{data}
\name{flat_titanic}
\alias{flat_titanic}
\title{Summaries of people and outcomes on titanic}
\format{
A data frame with 32 rows and 5 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
flat_titanic
}
\description{
A dataset containing blahhhhhhh
}
\keyword{datasets}
