% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{flat_titanic}
\alias{flat_titanic}
\title{Tallied characteristics}
\format{
A data frame with 32 rows and 5 variables:
\describe{
\item{class}{class (1st, 2nd, 3rd or Crew)}
\item{sex}{sex of passenger}
\item{age}{child or adult}
\item{survived}{survival outcome of passanger}
\item{freq}{tally of all in category}
...
}
}
\source{
R stats
}
\usage{
flat_titanic
}
\description{
A dataset tallying frequencies of titanic passenger characteristics based on Titanic data from base R.  Includes Crew
}
\keyword{datasets}
