% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{passengers}
\alias{passengers}
\title{Titanic test data.}
\format{
Data frame with columns
\describe{
\item{passenger_id}{Passenger ID}
\item{survived}{survival status}
\item{pclass}{Passenger Class}
\item{name}{Name}
\item{sex}{Sex}
\item{age}{Age}
\item{sib_sp}{Number of Siblings/Spouses Aboard}
\item{parch}{Number of Parents/Children Aboard}
\item{ticket}{Ticket Number}
\item{fare}{Passenger Fare}
\item{cabin}{Cabin}
\item{embarked}{Port of Embarkation}

...
}
}
\source{
titanic package, https://www.kaggle.com/c/titanic/data
}
\usage{
passengers
}
\description{
Titanic test data.
}
\keyword{datasets}
