% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexport.R, R/dataset_export.R
\docType{data}
\name{tidy_titanic}
\alias{tidy_titanic}
\title{Records for individuals on titanic.}
\format{
A data frame with 2201 rows and 5 variables:
\describe{
\item{age}{price, in US dollars}
\item{class}{weight of the diamond, in carats}
\item{sex}{weight of the diamond, in carats}
\item{survived}{weight of the diamond, in carats}
\item{id}{weight of the diamond, in carats}
...
}

A data frame with 2201 rows and 5 variables:
\describe{
\item{id}{an arbitrary id for each individual}
\item{class}{1st, 2nd, 3rd for each passenger}
\item{sex}{Male, Female for each passenger}
\item{age}{Child, Adult for each passenger}
\item{survived}{No, Yes for each passenger}
...
}
}
\source{
R stat

A version of the Titanic data from the datasets package
}
\usage{
tidy_titanic

tidy_titanic
}
\description{
A dataset ... blah blah

A dataset with a row for each individual on the titanic
}
\keyword{datasets}
