% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexport.R
\docType{data}
\name{tidy_titanic}
\alias{tidy_titanic}
\title{Records for individuals on titanic.}
\format{
A data frame with 2201 rows and 5 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
R stat
}
\usage{
tidy_titanic
}
\description{
A dataset ... blah blah
}
\keyword{datasets}
